// frontend_angular/src/app/api/model/port.ts
/**
 * Adherent
 * Adherent api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Port { 
    id?: number;
    portNumber: string;
    roomNumber?: number;
    switchID?: number;
}